+++ app/tests/test_order_full.py
@@ -5,7 +5,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 
 from tests.utils import url_reverse
 from routers.choices import OrderFilter
-from schemas.order_schema import OrderFullResponseSchema
+from entities.order.schemas.schema_order import OrderFullResponseSchema
 from routers.order_routers import ORDER_PAGE_SIZE
