+++ app/tests/test_order_payment.py
@@ -9,8 +9,8 @@ from entities.order.models_order import OrderModel
 from tests.utils import url_reverse
 from models.choices import StatusOrder
 from entities.order.models_order import OrderInSchema
-from schemas.order_schema import OrderPaymentSchema
-from schemas.payment_schema import CardSchema, PaymentType
+from entities.order.schemas.schema_order import OrderPaymentSchema
+from entities.order.schemas.schema_payment import CardSchema, PaymentType
 from services.order_service import OrderService
